fix: results out-of-order when `io_threads` is set larger than 1 (#93)
In previous version we use
[`submit`](https://docs.python.org/3/library/concurrent.futures.html#concurrent.futures.Executor.submit)
+
[`as_completed`](https://docs.python.org/3/library/concurrent.futures.html#concurrent.futures.as_completed)
to spawn tasks and collect results. However, `as_completed` doesn't
return results in the order of input futures. This led to a critical bug
that the results can be totally wrong.

This PR uses
[`map`](https://docs.python.org/3/library/concurrent.futures.html#concurrent.futures.Executor.map)
instead. Unit tests are also updated to check the correctness of
results.

I think all old versions should be yanked.

---------

Signed-off-by: Richard Chien <stdrc@outlook.com>