refactor!: Remove SubInterpreter and bump pyo3 to 0.24 (#125)
This PR will remove SubInterpreter and bump pyo3 to 0.24 as we discussed
before.

---

The new pyo3 doesn't allow us to hack the internal lock counter anymore.
There is no way for us to make `SubInterpreter` work. And keep pinning
on the old version of pyo3 makes us failed to work with python 3.13.

Thus we decided to remove SubInterpreter entirely.

This PR also completely removes the sandbox. Callers must ensure that
the input is trustworthy.

---------

Signed-off-by: Xuanwo <github@xuanwo.io>